Added region selection, changed styling, removed unnecessary button
Added a region selection that also sends the selected region to the database in firestore. This prompt is only asked once when starting the app. Changed the styling to have it look more simple but professional. Removed a button to go back to the city input that was unnecessary (button automatically exists when using navigationview)