Reusability changes
CalendarDay: Removed CalendarDay.swift as it was a copy with a different name to MoodEntry. Adapted files that used this model to now use MoodEntry.
EntryModel: removed EntryModel to replace with JournalEntry as they were copies of each other. Adapted all files that used EntryModel to now use JournalEntry as well.