Implemented the related article option
Added the logic to implement related articles related to recent frustrations from the user
FrustrationsView.swift: Modified the code related to the displaying of articles by using a reusable View.
ArticleListViewModel.swift: created a dedicated ViewModel to handle the connection to the dedicated service.
ArtcileListView.swift: created a reusable View that displays a list of articles based on a specific topic that can be chosen.
GoogleSearchService.swift: created a service that handles the searching of articles using the custom search api and a custom search engine using google cloud.
.gitignore: added the info.plist again as it was removed due to having no sensitive information which is now the case